move to resources, queries utils
