+++ localstack/services/events/provider.py
@@ -37,6 +37,7 @@ from localstack.services.moto import call_moto
 from localstack.services.plugins import ServiceLifecycleHook
 from localstack.utils.aws import aws_stack
 from localstack.utils.aws.message_forwarding import send_event_to_target
+from localstack.utils.collections import pick_attributes
 from localstack.utils.common import TMP_FILES, mkdir, save_file, truncate
 from localstack.utils.json import extract_jsonpath
 from localstack.utils.strings import long_uid, short_uid
@@ -81,7 +82,7 @@ class EventsProvider(EventsApi, ServiceLifecycleHook):
                 # TODO generate event matching aws in case no Input has been specified
                 event_str = target.get("Input") or "{}"
                 event = json.loads(event_str)
-                attr = aws_stack.get_events_target_attributes(target)
+                attr = pick_attributes(target, ["$.SqsParameters", "$.KinesisParameters"])
                 try:
                     send_event_to_target(arn, event, target_attributes=attr, target=target)
                 except Exception as e:
@@ -434,7 +435,10 @@ def process_events(event: Dict, targets: List[Dict]):
             changed_event = json.loads(target.get("Input"))
         try:
             send_event_to_target(
-                arn, changed_event, aws_stack.get_events_target_attributes(target), target=target
+                arn,
+                changed_event,
+                pick_attributes(target, ["$.SqsParameters", "$.KinesisParameters"]),
+                target=target,
             )
         except Exception as e:
             LOG.info(f"Unable to send event notification {truncate(event)} to target {target}: {e}")
