+++ localstack/utils/aws/aws_stack.py
@@ -10,7 +10,6 @@ from typing import Dict, Optional, Union
 from urllib.parse import urlparse
 
 from localstack.aws.accounts import get_aws_access_key_id, get_aws_account_id
-from localstack.utils.aws.arns import extract_region_from_arn, get_sqs_queue_url
 
 if sys.version_info >= (3, 8):
     pass
@@ -20,14 +19,12 @@ else:
 import boto3
 import botocore
 import botocore.config
-from botocore.utils import ArnParser
 
 from localstack import config
 from localstack.constants import (
     APPLICATION_AMZ_JSON_1_0,
     APPLICATION_AMZ_JSON_1_1,
     APPLICATION_X_WWW_FORM_URLENCODED,
-    AWS_REGION_US_EAST_1,
     ENV_DEV,
     HEADER_LOCALSTACK_ACCOUNT_ID,
     LOCALHOST,
@@ -37,12 +34,8 @@ from localstack.constants import (
     TEST_AWS_ACCESS_KEY_ID,
     TEST_AWS_SECRET_ACCESS_KEY,
 )
-from localstack.utils.aws.aws_models import KinesisStream
-from localstack.utils.collections import pick_attributes
-from localstack.utils.functions import run_safe
 from localstack.utils.http import make_http_request
-from localstack.utils.strings import get_random_hex, is_string, is_string_or_bytes, to_str
-from localstack.utils.sync import poll_condition
+from localstack.utils.strings import is_string, is_string_or_bytes, to_str
 
 # AWS environment variable names
 ENV_ACCESS_KEY = "AWS_ACCESS_KEY_ID"
@@ -67,11 +60,6 @@ BOTO_CLIENTS_CACHE = {}
 # Assume role loop seconds
 DEFAULT_TIMER_LOOP_SECONDS = 60 * 50
 
-# maps SQS queue ARNs to queue URLs
-SQS_ARN_TO_URL_CACHE = {}
-
-# List of parameters with additional event target parameters
-EVENT_TARGET_PARAMETERS = ["$.SqsParameters", "$.KinesisParameters"]
 
 # cached value used to determine the DNS status of the S3 hostname (whether it can be resolved properly)
 CACHE_S3_HOSTNAME_DNS_STATUS = None
@@ -496,54 +484,6 @@ def extract_access_key_id_from_auth_header(headers: Dict[str, str]) -> Optional[
             return access_id[0]
 
 
-# TODO: extract ARN utils into separate file!
-
-_arn_parser = ArnParser()
-
-
-def get_iam_role(resource, env=None):
-    env = get_environment(env)
-    return "role-%s" % resource
-
-
-def get_events_target_attributes(target):
-    return pick_attributes(target, EVENT_TARGET_PARAMETERS)
-
-
-def get_or_create_bucket(bucket_name: str, s3_client=None):
-    s3_client = s3_client or connect_to_service("s3")
-    try:
-        return s3_client.head_bucket(Bucket=bucket_name)
-    except Exception:
-        return create_s3_bucket(bucket_name, s3_client=s3_client)
-
-
-def create_s3_bucket(bucket_name: str, s3_client=None):
-    """Creates a bucket in the region that is associated with the current request
-    context, or with the given boto3 S3 client, if specified."""
-    s3_client = s3_client or connect_to_service("s3")
-    region = s3_client.meta.region_name
-    kwargs = {}
-    if region != AWS_REGION_US_EAST_1:
-        kwargs = {"CreateBucketConfiguration": {"LocationConstraint": region}}
-    return s3_client.create_bucket(Bucket=bucket_name, **kwargs)
-
-
-def create_sqs_queue(queue_name, env=None):
-    env = get_environment(env)
-    # queue
-    conn = connect_to_service("sqs", env=env)
-    return conn.create_queue(QueueName=queue_name)
-
-
-def sqs_receive_message(queue_arn):
-    region_name = extract_region_from_arn(queue_arn)
-    client = connect_to_service("sqs", region_name=region_name)
-    queue_url = get_sqs_queue_url(queue_arn)
-    response = client.receive_message(QueueUrl=queue_url)
-    return response
-
-
 def mock_aws_request_headers(
     service="dynamodb", region_name=None, access_key=None, internal=False
 ) -> Dict[str, str]:
@@ -581,6 +521,7 @@ def inject_region_into_auth_headers(region, headers):
         headers["Authorization"] = auth_header
 
 
+# TODO: is this still useful?
 def dynamodb_get_item_raw(request):
     headers = mock_aws_request_headers()
     headers["X-Amz-Target"] = "DynamoDB_20120810.GetItem"
@@ -593,244 +534,6 @@ def dynamodb_get_item_raw(request):
     new_item = new_item.text
     new_item = new_item and json.loads(new_item)
     return new_item
-
-
-def create_dynamodb_table(
-    table_name: str,
-    partition_key: str,
-    stream_view_type: str = None,
-    region_name: str = None,
-    client=None,
-    wait_for_active: bool = True,
-):
-    """Utility method to create a DynamoDB table"""
-
-    dynamodb = client or connect_to_service("dynamodb", region_name=region_name)
-    stream_spec = {"StreamEnabled": False}
-    key_schema = [{"AttributeName": partition_key, "KeyType": "HASH"}]
-    attr_defs = [{"AttributeName": partition_key, "AttributeType": "S"}]
-    if stream_view_type is not None:
-        stream_spec = {"StreamEnabled": True, "StreamViewType": stream_view_type}
-    table = None
-    try:
-        table = dynamodb.create_table(
-            TableName=table_name,
-            KeySchema=key_schema,
-            AttributeDefinitions=attr_defs,
-            BillingMode="PAY_PER_REQUEST",
-            StreamSpecification=stream_spec,
-        )
-    except Exception as e:
-        if "ResourceInUseException" in str(e):
-            # Table already exists -> return table reference
-            return connect_to_resource("dynamodb", region_name=region_name).Table(table_name)
-        if "AccessDeniedException" in str(e):
-            raise
-
-    def _is_active():
-        return dynamodb.describe_table(TableName=table_name)["Table"]["TableStatus"] == "ACTIVE"
-
-    if wait_for_active:
-        poll_condition(_is_active)
-
-    return table
-
-
-def get_apigateway_integration(api_id, method, path, env=None):
-    apigateway = connect_to_service(service_name="apigateway", client=True, env=env)
-
-    resources = apigateway.get_resources(restApiId=api_id, limit=100)
-    resource_id = None
-    for r in resources["items"]:
-        if r["path"] == path:
-            resource_id = r["id"]
-    if not resource_id:
-        raise Exception('Unable to find apigateway integration for path "%s"' % path)
-
-    integration = apigateway.get_integration(
-        restApiId=api_id, resourceId=resource_id, httpMethod=method
-    )
-    return integration
-
-
-def get_apigateway_resource_for_path(api_id, path, parent=None, resources=None):
-    if resources is None:
-        apigateway = connect_to_service(service_name="apigateway")
-        resources = apigateway.get_resources(restApiId=api_id, limit=100)
-    if not isinstance(path, list):
-        path = path.split("/")
-    if not path:
-        return parent
-    for resource in resources:
-        if resource["pathPart"] == path[0] and (not parent or parent["id"] == resource["parentId"]):
-            return get_apigateway_resource_for_path(
-                api_id, path[1:], parent=resource, resources=resources
-            )
-    return None
-
-
-def get_apigateway_path_for_resource(
-    api_id, resource_id, path_suffix="", resources=None, region_name=None
-):
-    if resources is None:
-        apigateway = connect_to_service(service_name="apigateway", region_name=region_name)
-        resources = apigateway.get_resources(restApiId=api_id, limit=100)["items"]
-    target_resource = list(filter(lambda res: res["id"] == resource_id, resources))[0]
-    path_part = target_resource.get("pathPart", "")
-    if path_suffix:
-        if path_part:
-            path_suffix = "%s/%s" % (path_part, path_suffix)
-    else:
-        path_suffix = path_part
-    parent_id = target_resource.get("parentId")
-    if not parent_id:
-        return "/%s" % path_suffix
-    return get_apigateway_path_for_resource(
-        api_id,
-        parent_id,
-        path_suffix=path_suffix,
-        resources=resources,
-        region_name=region_name,
-    )
-
-
-def create_api_gateway(
-    name,
-    description=None,
-    resources=None,
-    stage_name=None,
-    enabled_api_keys=None,
-    env=None,
-    usage_plan_name=None,
-    region_name=None,
-    auth_creator_func=None,  # function that receives an api_id and returns an authorizer_id
-    client=None,
-):
-    if enabled_api_keys is None:
-        enabled_api_keys = []
-    if not client:
-        client = connect_to_service("apigateway", env=env, region_name=region_name)
-    resources = resources or []
-    stage_name = stage_name or "testing"
-    usage_plan_name = usage_plan_name or "Basic Usage"
-    description = description or 'Test description for API "%s"' % name
-
-    LOG.info('Creating API resources under API Gateway "%s".', name)
-    api = client.create_rest_api(name=name, description=description)
-    api_id = api["id"]
-
-    auth_id = None
-    if auth_creator_func:
-        auth_id = auth_creator_func(api_id)
-
-    resources_list = client.get_resources(restApiId=api_id)
-    root_res_id = resources_list["items"][0]["id"]
-    # add API resources and methods
-    for path, methods in resources.items():
-        # create resources recursively
-        parent_id = root_res_id
-        for path_part in path.split("/"):
-            api_resource = client.create_resource(
-                restApiId=api_id, parentId=parent_id, pathPart=path_part
-            )
-            parent_id = api_resource["id"]
-        # add methods to the API resource
-        for method in methods:
-            kwargs = {"authorizerId": auth_id} if auth_id else {}
-            client.put_method(
-                restApiId=api_id,
-                resourceId=api_resource["id"],
-                httpMethod=method["httpMethod"],
-                authorizationType=method.get("authorizationType") or "NONE",
-                apiKeyRequired=method.get("apiKeyRequired") or False,
-                requestParameters=method.get("requestParameters") or {},
-                requestModels=method.get("requestModels") or {},
-                **kwargs,
-            )
-            # create integrations for this API resource/method
-            integrations = method["integrations"]
-            create_api_gateway_integrations(
-                api_id,
-                api_resource["id"],
-                method,
-                integrations,
-                env=env,
-                region_name=region_name,
-                client=client,
-            )
-    # deploy the API gateway
-    client.create_deployment(restApiId=api_id, stageName=stage_name)
-    return api
-
-
-def create_api_gateway_integrations(
-    api_id, resource_id, method, integrations=None, env=None, region_name=None, client=None
-):
-    if integrations is None:
-        integrations = []
-    if not client:
-        client = connect_to_service("apigateway", env=env, region_name=region_name)
-    for integration in integrations:
-        req_templates = integration.get("requestTemplates") or {}
-        res_templates = integration.get("responseTemplates") or {}
-        success_code = integration.get("successCode") or "200"
-        client_error_code = integration.get("clientErrorCode") or "400"
-        server_error_code = integration.get("serverErrorCode") or "500"
-        request_parameters = integration.get("requestParameters") or {}
-        # create integration
-        client.put_integration(
-            restApiId=api_id,
-            resourceId=resource_id,
-            httpMethod=method["httpMethod"],
-            integrationHttpMethod=method.get("integrationHttpMethod") or method["httpMethod"],
-            type=integration["type"],
-            uri=integration["uri"],
-            requestTemplates=req_templates,
-            requestParameters=request_parameters,
-        )
-        response_configs = [
-            {"pattern": "^2.*", "code": success_code, "res_templates": res_templates},
-            {"pattern": "^4.*", "code": client_error_code, "res_templates": {}},
-            {"pattern": "^5.*", "code": server_error_code, "res_templates": {}},
-        ]
-        # create response configs
-        for response_config in response_configs:
-            # create integration response
-            client.put_integration_response(
-                restApiId=api_id,
-                resourceId=resource_id,
-                httpMethod=method["httpMethod"],
-                statusCode=response_config["code"],
-                responseTemplates=response_config["res_templates"],
-                selectionPattern=response_config["pattern"],
-            )
-            # create method response
-            client.put_method_response(
-                restApiId=api_id,
-                resourceId=resource_id,
-                httpMethod=method["httpMethod"],
-                statusCode=response_config["code"],
-            )
-
-
-def get_opensearch_endpoint(domain_arn: str) -> str:
-    """
-    Get an OpenSearch cluster endpoint by describing the cluster associated with the domain_arn
-    :param domain_arn: ARN of the cluster.
-    :returns: cluster endpoint
-    :raises: ValueError if the domain_arn is malformed
-    """
-    region_name = extract_region_from_arn(domain_arn)
-    if region_name is None:
-        raise ValueError("unable to parse region from opensearch domain ARN")
-    opensearch_client = connect_to_service(service_name="opensearch", region_name=region_name)
-    domain_name = domain_arn.rpartition("/")[2]
-    info = opensearch_client.describe_domain(DomainName=domain_name)
-    base_domain = info["DomainStatus"]["Endpoint"]
-    # Add the URL scheme "http" if it's not set yet. https might not be enabled for all instances
-    # f.e. when the endpoint strategy is PORT or there is a custom opensearch/elasticsearch instance
-    endpoint = base_domain if base_domain.startswith("http") else f"http://{base_domain}"
-    return endpoint
 
 
 def get_search_db_connection(endpoint: str, region_name: str):
@@ -867,41 +570,3 @@ def get_search_db_connection(endpoint: str, region_name: str):
             http_auth=awsauth,
         )
     return OpenSearch(hosts=[endpoint], verify_certs=verify_certs, use_ssl=use_ssl)
-
-
-def create_kinesis_stream(stream_name, shards=1, env=None, delete=False):
-    env = get_environment(env)
-    stream = KinesisStream(id=stream_name, num_shards=shards)
-    conn = connect_to_service("kinesis", env=env)
-    stream.connect(conn)
-    if delete:
-        run_safe(lambda: stream.destroy(), print_error=False)
-    stream.create()
-    # Note: Returning the stream without awaiting its creation (via wait_for()) to avoid API call timeouts/retries.
-    return stream
-
-
-def kinesis_get_latest_records(stream_name, shard_id, count=10, env=None):
-    kinesis = connect_to_service("kinesis", env=env)
-    result = []
-    response = kinesis.get_shard_iterator(
-        StreamName=stream_name, ShardId=shard_id, ShardIteratorType="TRIM_HORIZON"
-    )
-    shard_iterator = response["ShardIterator"]
-    while shard_iterator:
-        records_response = kinesis.get_records(ShardIterator=shard_iterator)
-        records = records_response["Records"]
-        for record in records:
-            try:
-                record["Data"] = to_str(record["Data"])
-            except Exception:
-                pass
-        result.extend(records)
-        shard_iterator = records_response["NextShardIterator"] if records else False
-        while len(result) > count:
-            result.pop(0)
-    return result
-
-
-def get_trace_id():
-    return f"1-{get_random_hex(8)}-{get_random_hex(24)}"
