+++ tests/integration/test_apigateway.py
@@ -16,6 +16,8 @@ from requests.models import Response
 from requests.structures import CaseInsensitiveDict
 
 import localstack.utils.aws.arns
+import localstack.utils.aws.queries
+import localstack.utils.aws.resources
 from localstack import config
 from localstack.aws.accounts import get_aws_account_id
 from localstack.aws.handlers import cors
@@ -208,7 +210,9 @@ class TestAPIGateway:
 
     def test_api_gateway_kinesis_integration(self):
         # create target Kinesis stream
-        stream = aws_stack.create_kinesis_stream(self.TEST_STREAM_KINESIS_API_GW)
+        stream = localstack.utils.aws.resources.create_kinesis_stream(
+            self.TEST_STREAM_KINESIS_API_GW
+        )
         stream.wait_for()
 
         # create API Gateway and connect it to the target stream
@@ -249,7 +253,7 @@ class TestAPIGateway:
     def test_api_gateway_sqs_integration_with_event_source(self):
         # create target SQS stream
         queue_name = f"queue-{short_uid()}"
-        queue_url = aws_stack.create_sqs_queue(queue_name)["QueueUrl"]
+        queue_url = localstack.utils.aws.resources.create_sqs_queue(queue_name)["QueueUrl"]
 
         # create API Gateway and connect it to the target queue
         result = connect_api_gateway_to_sqs(
@@ -296,7 +300,7 @@ class TestAPIGateway:
     def test_api_gateway_sqs_integration(self):
         # create target SQS stream
         queue_name = f"queue-{short_uid()}"
-        aws_stack.create_sqs_queue(queue_name)
+        localstack.utils.aws.resources.create_sqs_queue(queue_name)
 
         # create API Gateway and connect it to the target queue
         result = connect_api_gateway_to_sqs(
@@ -317,7 +321,7 @@ class TestAPIGateway:
         result = requests.post(url, data=json.dumps(test_data))
         assert 200 == result.status_code
 
-        messages = aws_stack.sqs_receive_message(queue_name)["Messages"]
+        messages = localstack.utils.aws.queries.sqs_receive_message(queue_name)["Messages"]
         assert 1 == len(messages)
         assert test_data == json.loads(base64.b64decode(messages[0]["Body"]))
 
@@ -1715,7 +1719,7 @@ class TestAPIGateway:
         api_id, _, _ = create_rest_apigw(name=apigateway_name)
 
         try:
-            aws_stack.get_or_create_bucket(bucket_name)
+            localstack.utils.aws.resources.get_or_create_bucket(bucket_name)
             s3_client.put_object(
                 Bucket=bucket_name,
                 Key=object_name,
@@ -1878,7 +1882,7 @@ class TestAPIGateway:
                 },
             ]
         }
-        return aws_stack.create_api_gateway(
+        return localstack.utils.aws.resources.create_api_gateway(
             name=gateway_name, resources=resources, stage_name=self.TEST_STAGE_NAME
         )
 
@@ -1910,7 +1914,7 @@ class TestAPIGateway:
             }
             for method in methods
         ]
-        return aws_stack.create_api_gateway(
+        return localstack.utils.aws.resources.create_api_gateway(
             name=gateway_name, resources=resources, stage_name=self.TEST_STAGE_NAME
         )
 
@@ -2018,7 +2022,9 @@ class TestAPIGateway:
 
         kwargs = {}
         if integration_type == "AWS_PROXY":
-            aws_stack.create_dynamodb_table("MusicCollection", partition_key="id")
+            localstack.utils.aws.resources.create_dynamodb_table(
+                "MusicCollection", partition_key="id"
+            )
             kwargs[
                 "uri"
             ] = "arn:aws:apigateway:us-east-1:dynamodb:action/PutItem&Table=MusicCollection"
