+++ tests/integration/test_edge.py
@@ -9,6 +9,7 @@ import xmltodict
 from quart import request as quart_request
 from requests.models import Request as RequestsRequest
 
+import localstack.utils.aws.resources
 from localstack import config
 from localstack.aws.accounts import get_aws_account_id
 from localstack.constants import APPLICATION_JSON, HEADER_LOCALSTACK_EDGE_URL
@@ -83,7 +84,7 @@ class TestEdgeAPI:
     def _invoke_dynamodb_via_edge_go_sdk(self, edge_url):
         client = aws_stack.create_external_boto_client("dynamodb")
         table_name = f"t-{short_uid()}"
-        aws_stack.create_dynamodb_table(table_name, "id")
+        localstack.utils.aws.resources.create_dynamodb_table(table_name, "id")
 
         # emulate a request sent from the AWS Go SDK v2
         headers = {
