+++ tests/integration/test_multiregion.py
@@ -5,6 +5,7 @@ import unittest
 import requests
 
 import localstack.utils.aws.arns
+import localstack.utils.aws.queries
 from localstack import config
 from localstack.constants import PATH_USER_REQUEST
 from localstack.services.apigateway.helpers import connect_api_gateway_to_sqs
@@ -74,7 +75,7 @@ class TestMultiRegion(unittest.TestCase):
         test_data = {"foo": "bar"}
         result = requests.post(url, data=json.dumps(test_data))
         self.assertEqual(result.status_code, 200)
-        messages = aws_stack.sqs_receive_message(queue_arn)["Messages"]
+        messages = localstack.utils.aws.queries.sqs_receive_message(queue_arn)["Messages"]
         self.assertEqual(len(messages), 1)
         self.assertEqual(
             json.loads(to_str(base64.b64decode(to_str(messages[0]["Body"])))), test_data
